% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_species_area}
\alias{calculate_species_area}
\title{Species-Area (Accumulation) Data}
\usage{
calculate_species_area(abund_matrix)
}
\arguments{
\item{abund_matrix}{A site x species abundance data frame where the first
column is \code{site} and remaining columns are species abundances.}
}
\value{
A data frame with columns:
\describe{
\item{\code{Sites}}{Number of sites sampled.}
\item{\code{Richness}}{Mean cumulative species richness.}
\item{\code{SD}}{Standard deviation across permutations.}
}
}
\description{
Compute a species accumulation curve (SAR) using
\code{vegan::specaccum(method = "random")}.
}
\details{
Uses 100 random permutations by default to estimate mean richness
and its standard deviation as a function of the number of sites.
}
\examples{
\dontrun{
sar <- calculate_species_area(abund_matrix)
head(sar)
}
}
\seealso{
\code{\link{plot_species_area}}, \code{vegan::specaccum}
}
