% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointprocess.R
\name{.as_points_sf}
\alias{.as_points_sf}
\title{Build an sf POINT layer from a numeric matrix of XY}
\usage{
.as_points_sf(xy, crs)
}
\description{
Build an sf POINT layer from a numeric matrix of XY
}
\keyword{internal}
