% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrats_audit.R
\name{.attach_placement_audit}
\alias{.attach_placement_audit}
\title{Attach a placement_audit attribute to a quadrats sf object}
\usage{
.attach_placement_audit(quadrats, audit)
}
\arguments{
\item{quadrats}{sf object.}

\item{audit}{list.}
}
\value{
quadrats with normalised \code{placement_audit} attribute.
}
\description{
Attach a placement_audit attribute to a quadrats sf object
}
\keyword{internal}
