% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.has_cpp_geyer}
\alias{.has_cpp_geyer}
\title{Detect availability of the fast Geyer engine}
\usage{
.has_cpp_geyer()
}
\description{
Detect availability of the fast Geyer engine
}
\keyword{internal}
