% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.has_cpp_strauss}
\alias{.has_cpp_strauss}
\title{Detect availability of the fast Strauss engine}
\usage{
.has_cpp_strauss()
}
\description{
Detect availability of the fast Strauss engine
}
\keyword{internal}
