% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{.has_cpp_thomas}
\alias{.has_cpp_thomas}
\title{Detect availability of the fast Thomas engine}
\usage{
.has_cpp_thomas()
}
\description{
Detect availability of the fast Thomas engine
}
\keyword{internal}
