% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad.R
\name{generate_brokenstick_probabilities}
\alias{generate_brokenstick_probabilities}
\title{Broken-stick SAD probabilities}
\usage{
generate_brokenstick_probabilities(n_species)
}
\arguments{
\item{n_species}{Integer (>= 1). Number of species.}
}
\value{
Numeric vector of probabilities of length \code{n_species} summing to 1.
}
\description{
Simulates a broken-stick partition of the unit interval using \code{runif()}.
}
