% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad.R
\name{generate_lognormal_probabilities}
\alias{generate_lognormal_probabilities}
\title{Lognormal SAD probabilities (weights)}
\usage{
generate_lognormal_probabilities(n_species, meanlog = 0, sdlog = 1)
}
\arguments{
\item{n_species}{Integer (>= 1). Number of species.}

\item{meanlog, sdlog}{Passed to \code{\link[stats:Lognormal]{stats::rlnorm()}}.}
}
\value{
Numeric vector of probabilities of length \code{n_species} summing to 1.
}
\description{
Lognormal SAD probabilities (weights)
}
