% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usability.R
\name{theme_spesim}
\alias{theme_spesim}
\title{A consistent ggplot2 theme for spesim teaching plots}
\usage{
theme_spesim(base_size = 11)
}
\arguments{
\item{base_size}{Base font size.}
}
\description{
A consistent ggplot2 theme for spesim teaching plots
}
